+++ event-sourcing/event-repository/event-repository-jdbc/src/test/java/uk/gov/justice/services/eventsourcing/repository/jdbc/discovery/EventDiscoveryRepositoryIT.java
@@ -1,6 +1,6 @@
 package uk.gov.justice.services.eventsourcing.repository.jdbc.discovery;
 
-import static java.util.Optional.of;
+import static java.util.Optional.ofNullable;
 import static java.util.UUID.fromString;
 import static java.util.UUID.randomUUID;
 import static org.hamcrest.CoreMatchers.hasItems;
@@ -77,6 +77,33 @@ public class EventDiscoveryRepositoryIT {
     }
 
     @Test
+    public void shouldIgnoreRowsWithNullEventNumbers() throws Exception {
+
+        final DataSource eventStoreDataSource = new FrameworkTestDataSourceFactory().createEventStoreDataSource();
+
+        when(eventStoreDataSourceProvider.getDefaultDataSource()).thenReturn(eventStoreDataSource);
+
+        final UUID streamId = fromString("100bdbb4-da77-4f9f-b461-3980e3170001");
+        final int batchSize = 100;
+
+        final Event event_1 = anEvent(streamId, 1L, "event_1", 1L);
+        final Event event_2 = anEvent(streamId, 2L, "event_2", 2L);
+        final Event event_3 = anEvent(streamId, 3L, "event_3", 3L);
+        final Event event_4 = anEvent(streamId, 4L, "event_4", null);
+
+        insert(event_1);
+        insert(event_2);
+        insert(event_3);
+        insert(event_4);
+
+        final List<StreamPosition> streamPositions = eventDiscoveryRepository.getLatestStreamPositions(event_1.getId(), batchSize);
+
+        assertThat(streamPositions.size(), is(1));
+
+        assertThat(streamPositions.get(0), is(new StreamPosition(streamId, 3L)));
+    }
+
+    @Test
     public void shouldLimitResultsByBatchSize() throws Exception {
 
         final DataSource eventStoreDataSource = new FrameworkTestDataSourceFactory().createEventStoreDataSource();
@@ -103,6 +130,38 @@ public class EventDiscoveryRepositoryIT {
         assertThat(eventDiscoveryRepository.getLatestStreamPositions(event_2.getId(), batchSize_2).size(), is(batchSize_2));
     }
 
+    @Test
+    public void shouldFindMaxPositionInStreamForAllStreamsGreaterThanEventNumber() throws Exception {
+        final DataSource eventStoreDataSource = new FrameworkTestDataSourceFactory().createEventStoreDataSource();
+
+        when(eventStoreDataSourceProvider.getDefaultDataSource()).thenReturn(eventStoreDataSource);
+
+        final UUID streamId_1 = fromString("100bdbb4-da77-4f9f-b461-3980e3170001");
+        final UUID streamId_2 = fromString("200bdbb4-da77-4f9f-b461-3980e3170002");
+        final long maxPositionInStream_stream_1 = 11L;
+        final long maxPositionInStream_stream_2 = 22L;
+        final long eventNumberToSearchFrom = 2L;
+        final int batchSize = 100;
+
+        final Event event_1_1 = anEvent(streamId_1, 1L, "event_1_stream_1", 1L);
+        final Event event_1_2 = anEvent(streamId_1, maxPositionInStream_stream_1, "event_2_stream_2", eventNumberToSearchFrom);
+        final Event event_2_1 = anEvent(streamId_2, 1L, "event_1_stream_2", 3L);
+        final Event event_2_2 = anEvent(streamId_2, maxPositionInStream_stream_2, "event_2_stream_12", 4L);
+
+        insert(event_1_1);
+        insert(event_1_2);
+        insert(event_2_1);
+        insert(event_2_2);
+
+        final List<StreamPosition> streamPositions = eventDiscoveryRepository.getLatestStreamPositions(event_1_2.getId(), batchSize);
+
+        assertThat(streamPositions.size(), is(1));
+
+        assertThat(streamPositions, hasItems(
+                new StreamPosition(streamId_2, maxPositionInStream_stream_2)));
+
+    }
+
     private void insert(final Event event) throws Exception {
         final String sql = """
                 INSERT INTO event_log (
@@ -116,13 +175,15 @@ public class EventDiscoveryRepositoryIT {
                  VALUES (?, ?, ?, ?, ?, ?, ?)
                 """;
 
+
+        final Long eventNumber = event.getEventNumber().orElse(null);
         try(final Connection connection = eventStoreDataSource.getConnection();
             final PreparedStatement preparedStatement = connection.prepareStatement(sql)) {
             preparedStatement.setObject(1, event.getId());
             preparedStatement.setObject(2, event.getStreamId());
             preparedStatement.setLong(3, event.getPositionInStream());
             preparedStatement.setString(4, event.getName());
-            preparedStatement.setLong(5, event.getEventNumber().orElse(-1L));
+            preparedStatement.setObject(5, eventNumber);
             preparedStatement.setString(6, "some-metadata");
             preparedStatement.setString(7, "some-payload");
 
@@ -136,6 +197,6 @@ public class EventDiscoveryRepositoryIT {
                 streamId,
                 positionInStream, name, "some-metadata", "some-payload",
                 new UtcClock().now(),
-                of(eventNumber));
+                ofNullable(eventNumber));
     }
 }