+++ event-sourcing/event-source-api/src/main/java/uk/gov/justice/services/eventsourcing/discovery/EventSubscriptionDiscoveryBean.java
@@ -1,16 +1,13 @@
 package uk.gov.justice.services.eventsourcing.discovery;
 
-import static java.util.Collections.emptyList;
-import static java.util.List.of;
-import static java.util.UUID.randomUUID;
 import static javax.ejb.TransactionManagementType.CONTAINER;
 import static javax.transaction.Transactional.TxType.REQUIRES_NEW;
 
 import uk.gov.justice.services.eventsourcing.repository.jdbc.discovery.EventDiscoveryRepository;
 import uk.gov.justice.services.eventsourcing.repository.jdbc.discovery.StreamPosition;
 
-import java.util.Collections;
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 import javax.ejb.Stateless;
@@ -22,6 +19,8 @@ import javax.transaction.Transactional;
 @TransactionManagement(CONTAINER)
 public class EventSubscriptionDiscoveryBean {
 
+    private static final long ZEROTH_EVENT_NUMBER = 0L;
+
     @Inject
     private EventDiscoveryRepository eventDiscoveryRepository;
 
@@ -29,10 +28,15 @@ public class EventSubscriptionDiscoveryBean {
     private EventDiscoveryConfig eventDiscoveryConfig;
 
     @Transactional(REQUIRES_NEW)
-    public List<StreamPosition> discoverNewEvents(final UUID latestKnownEventId) {
+    public List<StreamPosition> discoverNewEvents(final Optional<UUID> latestKnownEventId) {
 
-        return eventDiscoveryRepository.getLatestStreamPositions(
-                latestKnownEventId,
-                eventDiscoveryConfig.getBatchSize());
+        final int batchSize = eventDiscoveryConfig.getBatchSize();
+        if(latestKnownEventId.isPresent()) {
+            return eventDiscoveryRepository.getLatestStreamPositions(
+                    latestKnownEventId.get(),
+                    batchSize);
+        }
+
+        return eventDiscoveryRepository.getLatestStreamPositions(ZEROTH_EVENT_NUMBER, batchSize);
     }
 }